♻️ [3/3] dynamic scheduler decoupling refactor (#3708)
