+++ services/director-v2/src/simcore_service_director_v2/api/routes/dynamic_services.py
@@ -28,7 +28,6 @@ from ...models.domains.dynamic_services import (
     RetrieveDataIn,
     RetrieveDataOutEnveloped,
 )
-from ...models.schemas.dynamic_services import SchedulerData
 from ...modules import projects_networks
 from ...modules.db.repositories.projects import ProjectsRepository
 from ...modules.db.repositories.projects_networks import ProjectsNetworksRepository
@@ -135,14 +134,13 @@ async def create_dynamic_service(
     if not await is_sidecar_running(
         service.node_uuid, dynamic_services_settings.DYNAMIC_SIDECAR
     ):
-        scheduler_data = SchedulerData.from_http_request(
+        await scheduler.add_service(
             service=service,
             simcore_service_labels=simcore_service_labels,
             port=dynamic_services_settings.DYNAMIC_SIDECAR.DYNAMIC_SIDECAR_PORT,
             request_dns=x_dynamic_sidecar_request_dns,
             request_scheme=x_dynamic_sidecar_request_scheme,
         )
-        await scheduler.add_service(scheduler_data)
 
     return cast(DynamicServiceGet, await scheduler.get_stack_status(service.node_uuid))
