+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/__init__.py
@@ -1,17 +1,7 @@
-from ._api import (
-    push_outputs,
-    remove_containers,
-    remove_sidecar_proxy_docker_networks_and_volumes,
-    save_state,
-)
 from ._task import DynamicSidecarsScheduler, setup_scheduler, shutdown_scheduler
 
 __all__: tuple[str, ...] = (
     "DynamicSidecarsScheduler",
-    "push_outputs",
-    "remove_containers",
-    "remove_sidecar_proxy_docker_networks_and_volumes",
-    "save_state",
     "setup_scheduler",
     "shutdown_scheduler",
 )
