+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_events_utils.py
@@ -6,9 +6,9 @@ from typing import Any, Deque, Final, Optional, cast
 
 from fastapi import FastAPI
 from models_library.projects_networks import ProjectsNetworks
+from models_library.projects_nodes import NodeID
 from models_library.projects_nodes_io import NodeIDStr
 from models_library.rabbitmq_messages import InstrumentationRabbitMessage
-from pydantic import AnyHttpUrl
 from servicelib.fastapi.long_running_tasks.client import (
     ProgressCallback,
     TaskClientResultError,
@@ -87,11 +87,22 @@ def are_all_user_services_containers_running(
     )
 
 
+def _get_scheduler_data(app: FastAPI, node_uuid: NodeID) -> SchedulerData:
+    dynamic_sidecars_scheduler: "DynamicSidecarsScheduler" = (
+        app.state.dynamic_sidecar_scheduler
+    )
+    # pylint: disable=protected-access
+    return dynamic_sidecars_scheduler._scheduler.get_scheduler_data(node_uuid)
+
+
 async def service_remove_containers(
+    app: FastAPI,
+    node_uuid: NodeID,
     dynamic_sidecar_client: DynamicSidecarClient,
-    scheduler_data: SchedulerData,
     progress_callback: Optional[ProgressCallback] = None,
 ) -> None:
+    scheduler_data: SchedulerData = _get_scheduler_data(app, node_uuid)
+
     try:
         await dynamic_sidecar_client.stop_service(
             scheduler_data.endpoint, progress_callback=progress_callback
@@ -108,31 +119,40 @@ async def service_remove_containers(
 
 
 async def service_save_state(
+    app: FastAPI,
+    node_uuid: NodeID,
     dynamic_sidecar_client: DynamicSidecarClient,
-    dynamic_sidecar_endpoint: AnyHttpUrl,
     progress_callback: Optional[ProgressCallback] = None,
 ) -> None:
+    scheduler_data: SchedulerData = _get_scheduler_data(app, node_uuid)
     await dynamic_sidecar_client.save_service_state(
-        dynamic_sidecar_endpoint, progress_callback=progress_callback
+        scheduler_data.endpoint, progress_callback=progress_callback
     )
 
 
 async def service_push_outputs(
+    app: FastAPI,
+    node_uuid: NodeID,
     dynamic_sidecar_client: DynamicSidecarClient,
-    dynamic_sidecar_endpoint: AnyHttpUrl,
     progress_callback: Optional[ProgressCallback] = None,
 ) -> None:
+    scheduler_data: SchedulerData = _get_scheduler_data(app, node_uuid)
     await dynamic_sidecar_client.push_service_output_ports(
-        dynamic_sidecar_endpoint, progress_callback=progress_callback
+        scheduler_data.endpoint, progress_callback=progress_callback
     )
 
 
 async def service_remove_sidecar_proxy_docker_networks_and_volumes(
     task_progress: TaskProgress,
     app: FastAPI,
-    scheduler_data: SchedulerData,
+    node_uuid: NodeID,
     dynamic_sidecar_settings: DynamicSidecarSettings,
+    set_were_state_and_outputs_saved: Optional[bool] = None,
 ) -> None:
+    scheduler_data: SchedulerData = _get_scheduler_data(app, node_uuid)
+
+    if set_were_state_and_outputs_saved is not None:
+        scheduler_data.dynamic_sidecar.were_state_and_outputs_saved = True
 
     # remove the 2 services
     task_progress.update(message="removing dynamic sidecar stack", percent=0.1)
@@ -192,7 +212,8 @@ async def service_remove_sidecar_proxy_docker_networks_and_volumes(
         ]
     )
 
-    await app.state.dynamic_sidecar_scheduler.remove_service_from_observation(
+    # pylint: disable=protected-access
+    await app.state.dynamic_sidecar_scheduler._scheduler.remove_service_from_observation(
         scheduler_data.node_uuid
     )
     scheduler_data.dynamic_sidecar.service_removal_state.mark_removed()
@@ -210,9 +231,10 @@ async def attempt_pod_removal_and_data_saving(
 
     async def _remove_containers_save_state_and_outputs() -> None:
         dynamic_sidecar_client: DynamicSidecarClient = get_dynamic_sidecar_client(app)
-        dynamic_sidecar_endpoint: AnyHttpUrl = scheduler_data.endpoint
 
-        await service_remove_containers(dynamic_sidecar_client, scheduler_data)
+        await service_remove_containers(
+            app, scheduler_data.node_uuid, dynamic_sidecar_client
+        )
 
         # only try to save the status if :
         # - it is requested to save the state
@@ -227,7 +249,7 @@ async def attempt_pod_removal_and_data_saving(
             try:
                 tasks = [
                     service_push_outputs(
-                        dynamic_sidecar_client, dynamic_sidecar_endpoint
+                        app, scheduler_data.node_uuid, dynamic_sidecar_client
                     )
                 ]
 
@@ -236,7 +258,7 @@ async def attempt_pod_removal_and_data_saving(
                 if not app_settings.DIRECTOR_V2_DEV_FEATURE_R_CLONE_MOUNTS_ENABLED:
                     tasks.append(
                         service_save_state(
-                            dynamic_sidecar_client, dynamic_sidecar_endpoint
+                            app, scheduler_data.node_uuid, dynamic_sidecar_client
                         )
                     )
 
@@ -283,7 +305,7 @@ async def attempt_pod_removal_and_data_saving(
         await _remove_containers_save_state_and_outputs()
 
     await service_remove_sidecar_proxy_docker_networks_and_volumes(
-        TaskProgress.create(), app, scheduler_data, dynamic_sidecar_settings
+        TaskProgress.create(), app, scheduler_data.node_uuid, dynamic_sidecar_settings
     )
 
     # instrumentation
