+++ services/director-v2/tests/unit/conftest.py
@@ -96,12 +96,12 @@ def request_scheme() -> str:
 
 @pytest.fixture
 def scheduler_data_from_http_request(
-    run_id: RunID,
     dynamic_service_create: DynamicServiceCreate,
     simcore_service_labels: SimcoreServiceLabels,
     dynamic_sidecar_port: int,
     request_dns: str,
     request_scheme: str,
+    run_id: RunID,
 ) -> SchedulerData:
     return SchedulerData.from_http_request(
         service=dynamic_service_create,
@@ -412,7 +412,7 @@ def caplog_info_level(caplog: LogCaptureFixture) -> Iterable[LogCaptureFixture]:
 def mock_docker_api(mocker: MockerFixture) -> None:
     module_base = "simcore_service_director_v2.modules.dynamic_sidecar.scheduler"
     mocker.patch(
-        f"{module_base}._task.get_dynamic_sidecars_to_observe",
+        f"{module_base}._core._scheduler.get_dynamic_sidecars_to_observe",
         autospec=True,
         return_value=[],
     )
@@ -422,7 +422,7 @@ def mock_docker_api(mocker: MockerFixture) -> None:
         return_value=True,
     )
     mocker.patch(
-        f"{module_base}._task.get_dynamic_sidecar_state",
+        f"{module_base}._core._scheduler.get_dynamic_sidecar_state",
         return_value=(ServiceState.PENDING, ""),
     )
