+++ services/director-v2/tests/unit/test_api_route_dynamic_scheduler.py
@@ -8,10 +8,14 @@ from typing import AsyncIterator
 import pytest
 import respx
 from fastapi import status
+from models_library.service_settings_labels import SimcoreServiceLabels
 from pytest import MonkeyPatch
 from pytest_mock.plugin import MockerFixture
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from requests import Response
+from simcore_service_director_v2.models.domains.dynamic_services import (
+    DynamicServiceCreate,
+)
 from simcore_service_director_v2.models.schemas.dynamic_services.scheduler import (
     SchedulerData,
 )
@@ -73,10 +77,24 @@ async def mock_sidecar_api(scheduler_data: SchedulerData) -> AsyncIterator[None]
 
 @pytest.fixture
 async def observed_service(
-    dynamic_sidecar_scheduler: DynamicSidecarsScheduler, scheduler_data: SchedulerData
+    dynamic_sidecar_scheduler: DynamicSidecarsScheduler,
+    dynamic_service_create: DynamicServiceCreate,
+    simcore_service_labels: SimcoreServiceLabels,
+    dynamic_sidecar_port: int,
+    request_dns: str,
+    request_scheme: str,
 ) -> SchedulerData:
-    await dynamic_sidecar_scheduler.add_service(scheduler_data)
-    return scheduler_data
+    await dynamic_sidecar_scheduler.add_service(
+        dynamic_service_create,
+        simcore_service_labels,
+        dynamic_sidecar_port,
+        request_dns,
+        request_scheme,
+    )
+    # pylint:disable=protected-access
+    return dynamic_sidecar_scheduler._scheduler.get_scheduler_data(
+        dynamic_service_create.node_uuid
+    )
 
 
 @pytest.fixture
