+++ services/director-v2/tests/unit/test_modules_dynamic_sidecar_scheduler_task.py
@@ -224,9 +224,9 @@ async def test_skip_observation_cycle_after_error(
     # add a task, emulate an error make sure no observation cycle is
     # being triggered again
     assert mocked_dynamic_scheduler_events.count == 0
-    await scheduler.add_service(scheduler_data)
+    await scheduler._scheduler._add_service(scheduler_data)
     # check it is being tracked
-    assert scheduler_data.node_uuid in scheduler._inverse_search_mapping
+    assert scheduler_data.node_uuid in scheduler._scheduler._inverse_search_mapping
 
     # ensure observation cycle triggers a lot
     await asyncio.sleep(SCHEDULER_INTERVAL_SECONDS * 10)
@@ -237,8 +237,15 @@ async def test_skip_observation_cycle_after_error(
     # check if service was properly removed or is still kept for manual interventions
     if error_raised_by_saving_state:
         if use_case.outcome_service_removed:
-            assert scheduler_data.node_uuid not in scheduler._inverse_search_mapping
+            assert (
+                scheduler_data.node_uuid
+                not in scheduler._scheduler._inverse_search_mapping
+            )
         else:
-            assert scheduler_data.node_uuid in scheduler._inverse_search_mapping
+            assert (
+                scheduler_data.node_uuid in scheduler._scheduler._inverse_search_mapping
+            )
     else:
-        assert scheduler_data.node_uuid not in scheduler._inverse_search_mapping
+        assert (
+            scheduler_data.node_uuid not in scheduler._scheduler._inverse_search_mapping
+        )
